Switch from uuid4() to an sha1 hash of the upload
This simple change effectively makes uploading idempotent.  Because
the id of the document is now based on the contents of the data being
uploaded, uploading the same diff results in the same hash, and thus
same document id.  So basically it's idempotent because the second
persist will be a readonly noop (other than the read to see if it's
there in the first place).