Initial import of a [barely working] client
Currently Git is the only scm supported, and it's only been tested
with one specific diff, aka it's undoubtely very brittle.  Current
usage is to pipe the output of git diff to the client.  The client
isn't executable thus you need to pass the client to the python
interpreter.  The file hasn't been made +x because when it's actually
used the build will be creating [via entry_points] a script that
essentially calls pubdiff.client.main().

Other changes in this commit:

1. Reduced the H1 font size
2. Fixed css float where it wasn't getting cleared per diff
3. Moved the name of the file on top of the respective code block
4. Changed the upload service to return a dict with a url key