Properly handle unstaged changes
Previously if you had unstaged changes (aka changes not yet on the
index) the lookup would fail, because the hash specified in the diff
isn't something "git show" can fetch.  When this happens, compare the
hash with the hash of the file on disk.  If they are the same, we know
we're dealing with local changes.

Also updated a few regexes to use raw strings, and added some
conditional debugging statements.