Remove fancy formatting
The client side (browser) rendering was unacceptable with large diffs.
Turns out the cost of applying syntax highlighting client side (which
requires massive updates to the DOM) is just too expensive.  For now
regress back to a very crude diff with simple colors.  It's not as
fancy, but should have very good performance.